requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
dateparser==1.1.8
decorator==5.1.1
dateparser==1.1.8
Flask==2.2.3
Flask-HTTPAuth==4.7.0
frozenlist==1.3.3
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.2
openai==0.27.2
packaging==23.0
pathy==0.10.1
preshed==3.0.8
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
ratelim==0.1.6
regex==2023.3.23
requests==2.28.2
six==1.16.0
smart-open==6.3.0
srsly==2.4.6
thinc==8.1.9
tqdm==4.65.0
typer==0.7.0
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
wasabi==1.1.1
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
